fix: re-encode session buckets, gate fetch dialing, and report wire gaps
Session buckets now percent-encode `/` and `%` so two workspaces can
no longer share one, and buckets written under the old folding key are
refiled at startup from the workspace their transcripts record.

The fetch tool dials through a resolver that refuses private
addresses, follows redirects only with a fresh verdict, and honors no
env proxy while the gate is on. Spill directories and the settings
file are written user-private; `write_private` suffixes its temp with
the pid so two writers cannot clobber one.

Wire decoders now name the fields an endpoint owed but did not send —
only on the frames that carry them — and a truncated refusal body can
no longer halve a number `overflow_limit` would read. A subagent that
runs past its stop grace says what spend went uncounted, a compaction
that only pruned notices says so, and a `leave`/`resume` cycle around
an external editor reinstates the terminal-restoring panic hook.

+++ crates/cli/src/tui/panel.rs
@@ -742,7 +742,7 @@ mod tests {
         // a wrapped value may land the caret anywhere, but never past the
         // text actually on its row.
         assert_eq!(
-            crate::render::visible_width(&rows[r as usize]) as usize,
+            crate::render::visible_width(&rows[r as usize]),
             col as usize,
             "the caret is at the end of its own row"
         );
